🗃️ Adds Pause persistence
This commit creates the Pause.kt entity, which is stored in
PauseRepository.kt. The tests for Pause functions are in PauseTest.kt.
For PauseRepository.kt, PauseRepositoryTest.kt.

Both classes had similar behavior, and couldn't be run in parallel as
ObjectBox must be instantiated only once throughout the whole JVM
duration. A change to how the ObjectBox is used in tests, leveraging
Kotest's Listeners was made, tidying ObjectBox testing up.

Part of #86

Signed-off-by: Leonardo Colman Lopes <dev@leonardo.colman.com.br>